IP-17 #comment I have uploaded to Git a risk matrix, this is the prototype and I shall be editing it further throughout the week

+++ src/main/java/com/qa/ims/persistence/dao/ItemsDAO.java
@@ -22,7 +22,7 @@ public class ItemsDAO implements Dao<Items> {
 	public Items modelFromResultSet(ResultSet resultSet) throws SQLException {
 		Long itemID = resultSet.getLong("ItemID");
 		String orderItem = resultSet.getString("order_item");
-		Double price = resultSet.getFloat("price");
+		Double price = resultSet.getLong("price");
 		return new Items(itemID, orderItem, price);
 	}
